requirements.txt
aiohttp==2.2.5
aiopg==0.13.0
aioredis==1.0.0
appnope==0.1.0
async_generator==1.9
async-timeout==1.2.1
attrdict==2.0.0
autobahn==17.7.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.1.2
enum34==1.1.6
idna==2.6
inflection==0.3.1
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jsonschema==2.6.0
Markdown==2.4
multidict==3.1.3
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
python-jsonschema-objects==0.2.2
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.13
telepot==12.3
traitlets==4.3.2
txaio==2.8.1
urllib3==1.22
wcwidth==0.1.7
websockets==3.3
yarl==0.12.0
aioredis==1.0.0
pytest==3.3.2
pytest-asyncio==0.8.0
pytest-xdist==1.22.2
hypothesis
raven==6.6.0
raven_aiohttp==0.6.0
